Fix import cycle, get tests passing and binary compiling
